Add vCD 9.1 schemas and bindings.
This commit adds the schemas for the REST and Swagger-based APIs, and for
object extensibility. This commit contains the code that originally lived
in the api-schemas folder of the vmware/vcd-ext-sdk project plus the
bindings package from the ui/api-client folder of that same project. This
reorganization of code should allow for a more appropriate versioning and
release cycle for public interfaces to vCloud Director.

There were a few structural changes to shuffle the existing Maven projects
around into a more sensible hierarchy; separating schemas from bindings,
and separating bindings by language. Additionally, the Maven artifacts
were adjusted for consistency and for eventual publishing.

Testing Done:
Starting with an empty Maven local repository, verified that a mvn install
completed successfully. Visually inspected the resulting bindings JARs,
comparing the sizes to the output of those same artifacts from a
vmware/vcd-ext-sdk build. The sizes of the compiled source matched,
providing some amount of comfort that the produced bits are generally the
same. Also inspected the distribution of the typescript output to ensure
that it was structured as an appropriate NPM package.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>