Fix deploy issues to Sonatype staging repository.
This change adds the <url> element to the parent POM (to accompany the
organization URL which by itself is insufficient). It also reenables
Javadoc generation on the generated sources for api-bindings and for
extensibility-bindings. The failOnError flag needs to be set to false for
this, since the generated code produces invalid javadocs. Since Sonatype
deployment requires Javadocs for any JAR containing classes, I decided
that invalid Javadoc was better than just generating a dummy Javadoc
artifact to trick Sonatype and give the illusion of Javadocs.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>