[ui] Support for publishing to NPM
This switches the UI to NPM 6.x (currently the flavour of the month
for package management in UI) and adds publishing support in the
deploy phase.

The flow is:
  mvn deploy
  |_ release profile
    |_ frontend plugin
      |_ npm run mvn:deploy
        |_ .npmrc.dist used as .npmrc for npm publish target/dist
          |_ .npmrc checked for registry
            |_ NPM_API_TOKEN env var used for auth token

This currently is set to dry run only.  Once proven to work, we
will remove the dry run option.

Testing done: local mvn clean install deploy
Signed-off-by: David Byard <dbyard@vmware.com>