Consume new version of bindings libraries. (#9)
Version 0.9.2 of the bindings libraries address an issue in generating the
intex.ts file for the @vcd/bindings package in certain conditions (which
includes building in the Travis CI environment for publishing).

Versions in this repository have been updated in preparation for a patch
release.

Testing done:
Executed a build of this repository with updated dependencies. Ran another
build without cleaning (a Travis CI equivalent). Manually inspected
intex.ts and verified correct structure.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>